perf: doesn't depend on as much shit, doesn't reload from disk each time

+++ src/defs.jl
@@ -9,8 +9,8 @@ const Strip = SubArray{
     2,
     Matrix{PIXEL},
     Tuple{
-        UnitRange{Int64},
-        UnitRange{Int64}
+        UnitRange{Int},
+        UnitRange{Int}
     },
     false
 }
